HREPTWO-2364: make sure the preview is set to null when closed; additional guard to avoid problem in the first place added too.
svn path=/hippo-ecm/trunk/; revision=17408

+++ cms/addon/cms/edit/src/main/java/org/hippoecm/frontend/plugins/cms/edit/CmsEditor.java
@@ -105,25 +105,27 @@ class CmsEditor implements IEditor {
     }
 
     public void close() throws EditorException {
-        List<IEditorFilter> filters = context.getServices(context.getReference(this).getServiceId(),
-                IEditorFilter.class);
-        IdentityHashMap<IEditorFilter, Object> filterContexts = new IdentityHashMap<IEditorFilter, Object>();
-        for (IEditorFilter filter : filters) {
-            Object filterContext = filter.preClose();
-            if (filterContext == null) {
-                throw new EditorException("Close operation cancelled by filter");
+        if(context.getReference(this) != null) {
+            List<IEditorFilter> filters = context.getServices(context.getReference(this).getServiceId(),
+                    IEditorFilter.class);
+            IdentityHashMap<IEditorFilter, Object> filterContexts = new IdentityHashMap<IEditorFilter, Object>();
+            for (IEditorFilter filter : filters) {
+                Object filterContext = filter.preClose();
+                if (filterContext == null) {
+                    throw new EditorException("Close operation cancelled by filter");
+                }
+                filterContexts.put(filter, filterContext);
             }
-            filterContexts.put(filter, filterContext);
-        }
 
-        String renderId = context.getReference(renderer).getServiceId();
-        context.unregisterService(focusListener, renderId);
-        context.unregisterService(this, renderId);
+            String renderId = context.getReference(renderer).getServiceId();
+            context.unregisterService(focusListener, renderId);
+            context.unregisterService(this, renderId);
 
-        cluster.stop();
-        modelService.destroy();
-        for (IEditorFilter filter : filters) {
-            filter.postClose(filterContexts.get(filter));
+            cluster.stop();
+            modelService.destroy();
+            for (IEditorFilter filter : filters) {
+                filter.postClose(filterContexts.get(filter));
+            }
         }
         context.unregisterService(this, cluster.getClusterConfig().getString("editor.id"));
         manager.unregister(this);
